+++ Data Structure and algorithms/binarySearch.cpp
