+++ Java/star.java
