+++ Python/gcd.py
